// src/app/shared/interfaces/medium.interface.ts
export interface MediumInterface {
  id: number;
  img: string;
  statusIcon: string;
  statusText: string;
  name: string;
  skills: string;
  ratingIcon: string;
  priceFirst: string;
  priceSecond: string;
  priceThird: string;
  btn: string;
  btnStyle: string;
  specialty: string;
  about: string;
  isLiked: boolean;
}